+++ tests/src/acceptNft.test.ts
@@ -1,14 +1,15 @@
 import { expect } from "chai";
 import { getApiConnection } from "./substrate/substrate-api";
-import { acceptNft } from "./util/acceptNft";
-import { createCollection } from "./util/createCollection";
-import { isNftChildOfAnother } from "./util/getChildren";
-import { Nft } from "./util/makeGeneralNftOwner";
-import { mintNft } from "./util/mintNft";
-import { sendNft } from "./util/sendNft";
-import { expectTxFailure } from "./util/txResult";
+import {
+    createCollection,
+    mintNft,
+    sendNft,
+    acceptNft
+} from "./util/tx";
+import { NftIdTuple } from "./util/fetch";
+import { isNftChildOfAnother, expectTxFailure } from "./util/helpers";
 
-describe("Integration test: send NFT", () => {
+describe("Integration test: accept NFT", () => {
     let api: any;
     before(async () => { api = await getApiConnection(); });
 
@@ -35,16 +36,16 @@ describe("Integration test: send NFT", () => {
         const parentNftId = await mintNft(api, alice, ownerAlice, aliceCollectionId, "parent-nft-metadata");
         const childNftId = await mintNft(api, bob, ownerBob, bobCollectionId, "child-nft-metadata");
 
-        const newOwnerNFT: Nft = [aliceCollectionId, parentNftId];
+        const newOwnerNFT: NftIdTuple = [aliceCollectionId, parentNftId];
 
         await sendNft(api, "pending", ownerBob, bobCollectionId, childNftId, newOwnerNFT);
         await acceptNft(api, alice, bobCollectionId, childNftId, newOwnerNFT);
 
-        const isChild = await isNftChildOfAnother(api, [bobCollectionId, childNftId], newOwnerNFT);
+        const isChild = await isNftChildOfAnother(api, bobCollectionId, childNftId, newOwnerNFT);
         expect(isChild).to.be.true;
     });
 
-    it("[Negative] unable to accept NFT by a not-an-owner", async () => {
+    it("[negative] unable to accept NFT by a not-an-owner", async () => {
         const ownerAlice = alice;
         const ownerBob = bob;
 
@@ -54,18 +55,18 @@ describe("Integration test: send NFT", () => {
         const parentNftId = await mintNft(api, alice, ownerAlice, aliceCollectionId, "parent-nft-metadata");
         const childNftId = await mintNft(api, bob, ownerBob, bobCollectionId, "child-nft-metadata");
 
-        const newOwnerNFT: Nft = [aliceCollectionId, parentNftId];
+        const newOwnerNFT: NftIdTuple = [aliceCollectionId, parentNftId];
 
         await sendNft(api, "pending", ownerBob, bobCollectionId, childNftId, newOwnerNFT);
         const tx = acceptNft(api, bob, bobCollectionId, childNftId, newOwnerNFT);
 
         await expectTxFailure(/NoPermission/, tx);
 
-        const isChild = await isNftChildOfAnother(api, [bobCollectionId, childNftId], newOwnerNFT);
+        const isChild = await isNftChildOfAnother(api, bobCollectionId, childNftId, newOwnerNFT);
         expect(isChild).to.be.false;
     });
 
-    it("[Negative] unable to accept non-existing NFT", async () => {
+    it("[negative] unable to accept non-existing NFT", async () => {
         const collectionId = 0;
         const maxNftId = 0xFFFFFFFF;
 
@@ -74,14 +75,14 @@ describe("Integration test: send NFT", () => {
 
         const parentNftId = await mintNft(api, alice, owner, aliceCollectionId, "parent-nft-metadata");
 
-        const newOwnerNFT: Nft = [aliceCollectionId, parentNftId];
+        const newOwnerNFT: NftIdTuple = [aliceCollectionId, parentNftId];
 
         const tx = acceptNft(api, alice, collectionId, maxNftId, newOwnerNFT);
 
         await expectTxFailure(/NoAvailableNftId/, tx);
     });
 
-    it("[Negative] unable to accept NFT which is not sent", async () => {
+    it("[negative] unable to accept NFT which is not sent", async () => {
         const ownerAlice = alice;
         const ownerBob = bob;
 
@@ -91,13 +92,13 @@ describe("Integration test: send NFT", () => {
         const parentNftId = await mintNft(api, alice, ownerAlice, aliceCollectionId, "parent-nft-metadata");
         const childNftId = await mintNft(api, bob, ownerBob, bobCollectionId, "child-nft-metadata");
 
-        const newOwnerNFT: Nft = [aliceCollectionId, parentNftId];
+        const newOwnerNFT: NftIdTuple = [aliceCollectionId, parentNftId];
 
         const tx = acceptNft(api, alice, bobCollectionId, childNftId, newOwnerNFT);
 
         await expectTxFailure(/NoPermission/, tx);
 
-        const isChild = await isNftChildOfAnother(api, [bobCollectionId, childNftId], newOwnerNFT);
+        const isChild = await isNftChildOfAnother(api, bobCollectionId, childNftId, newOwnerNFT);
         expect(isChild).to.be.false;
     });
