+++ tests/src/addTheme.test.ts
@@ -1,8 +1,7 @@
 import { expect } from 'chai';
 import { getApiConnection } from './substrate/substrate-api';
-import { addTheme } from "./util/addTheme";
-import { createBase } from "./util/createBase";
-import { expectTxFailure } from './util/txResult';
+import { createBase, addTheme } from "./util/tx";
+import { expectTxFailure } from './util/helpers';
 
 describe("Integration test: add Theme to Base", () => {
     let api: any;
@@ -11,7 +10,7 @@ describe("Integration test: add Theme to Base", () => {
     const alice = "//Alice";
     const bob = "//Bob";
 
-    it("Add default theme", async () => {
+    it("add default theme", async () => {
         const baseId = await createBase(api, alice, "default-themed-base", "DTBase", []);
         await addTheme(api, alice, baseId, {
             name: "default",
@@ -28,7 +27,7 @@ describe("Integration test: add Theme to Base", () => {
         });
     });
 
-    it("Add default theme and a custom one", async () => {
+    it("add default theme and a custom one", async () => {
         const baseId = await createBase(api, alice, "2-themed-base", "2TBase", []);
         await addTheme(api, alice, baseId, {
             name: "default",
@@ -54,7 +53,7 @@ describe("Integration test: add Theme to Base", () => {
         })
     });
 
-    it("[Negative] Unable to add theme to non-existing base", async () => {
+    it("[negative] unable to add theme to non-existing base", async () => {
         const maxBaseId = 0xFFFFFFFF;
         const tx = addTheme(api, alice, maxBaseId, {
             name: "default",
@@ -64,7 +63,7 @@ describe("Integration test: add Theme to Base", () => {
         await expectTxFailure(/BaseDoesntExist/, tx);
     });
 
-    it("[Negative] Unable to add custom theme if no default theme", async () => {
+    it("[negative] unable to add custom theme if no default theme", async () => {
         const baseId = await createBase(api, alice, "no-default-themed-base", "NDTBase", []);
         const tx = addTheme(api, alice, baseId, {
             name: "custom-theme",
@@ -74,7 +73,7 @@ describe("Integration test: add Theme to Base", () => {
         await expectTxFailure(/NeedsDefaultThemeFirst/, tx);
     });
 
-    it("[Negative] Unable to add theme by a not-an-owner", async () => {
+    it("[negative] unable to add theme by a not-an-owner", async () => {
         const baseId = await createBase(api, alice, "no-default-themed-base", "NDTBase", []);
         const tx = addTheme(api, bob, baseId, {
             name: "default",
