+++ tests/src/createCollection.test.ts
@@ -1,5 +1,5 @@
 import { getApiConnection } from './substrate/substrate-api';
-import {createCollection} from './util/createCollection';
+import { createCollection } from './util/tx';
 
 describe('Integration test: create new collection', () => {
     let api: any;
@@ -7,11 +7,11 @@ describe('Integration test: create new collection', () => {
 
     const alice = '//Alice';
 
-    it('Create NFT collection', async () => {
+    it('create NFT collection', async () => {
         await createCollection(api, alice, 'test-metadata', 42, 'test-symbol');
     });
 
-    it('Create NFT collection without token limit', async () => {
+    it('create NFT collection without token limit', async () => {
         await createCollection(api, alice, 'no-limit-metadata', null, 'no-limit-symbol');
     });
