+++ tests/src/sendNft.test.ts
@@ -1,11 +1,8 @@
 import { expect } from "chai";
 import { getApiConnection } from "./substrate/substrate-api";
-import { createCollection } from "./util/createCollection";
-import { getChildren, isNftChildOfAnother } from "./util/getChildren";
-import { Nft } from "./util/makeGeneralNftOwner";
-import { mintNft } from "./util/mintNft";
-import { sendNft } from "./util/sendNft";
-import { expectTxFailure } from "./util/txResult";
+import { createCollection, mintNft, sendNft } from "./util/tx";
+import { NftIdTuple } from "./util/fetch";
+import { isNftChildOfAnother, expectTxFailure } from "./util/helpers";
 
 describe("Integration test: send NFT", () => {
     let api: any;
@@ -38,7 +35,7 @@ describe("Integration test: send NFT", () => {
         await sendNft(api, "sent", originalOwnerUri, collectionId, nftId, newOwnerUri);
     });
 
-    it("[Negative] unable to send non-existing NFT", async () => {
+    it("[negative] unable to send non-existing NFT", async () => {
         const originalOwnerUri = alice;
         const newOwnerUri = bob;
 
@@ -48,7 +45,7 @@ describe("Integration test: send NFT", () => {
         await expectTxFailure(/NoAvailableNftId/, tx);
     });
 
-    it("[Negative] unable to send NFT by a not-an-owner", async () => {
+    it("[negative] unable to send NFT by a not-an-owner", async () => {
         const originalOwnerUri = alice;
         const newOwnerUri = bob;
 
@@ -68,7 +65,7 @@ describe("Integration test: send NFT", () => {
         const parentNftId = await mintNft(api, alice, originalOwnerUri, collectionId, "parent-nft-metadata");
         const childNftId = await mintNft(api, alice, originalOwnerUri, collectionId, "child-nft-metadata");
 
-        const newOwnerNFT: Nft = [collectionId, parentNftId];
+        const newOwnerNFT: NftIdTuple = [collectionId, parentNftId];
 
         await sendNft(api, "sent", alice, collectionId, childNftId, newOwnerNFT);
 
@@ -76,7 +73,7 @@ describe("Integration test: send NFT", () => {
         expect(isChild).to.be.true;
     });
 
-    it("[Negative] send non-existing NFT to another NFT", async () => {
+    it("[negative] send non-existing NFT to another NFT", async () => {
         const originalOwnerUri = alice;
 
         const collectionId = await createTestCollection(alice);
@@ -84,7 +81,7 @@ describe("Integration test: send NFT", () => {
         const parentNftId = await mintNft(api, alice, originalOwnerUri, collectionId, "parent-nft-metadata");
         const childNftId = maxNftId;
 
-        const newOwnerNFT: Nft = [collectionId, parentNftId];
+        const newOwnerNFT: NftIdTuple = [collectionId, parentNftId];
 
         const tx = sendNft(api, "sent", alice, collectionId, childNftId, newOwnerNFT);
 
@@ -105,7 +102,7 @@ describe("Integration test: send NFT", () => {
         const parentNftId = await mintNft(api, author, originalOwnerUri, collectionId, "parent-nft-metadata");
         const childNftId = await mintNft(api, author, originalOwnerUri, collectionId, "child-nft-metadata");
 
-        const newOwnerNFT: Nft = [collectionId, parentNftId];
+        const newOwnerNFT: NftIdTuple = [collectionId, parentNftId];
 
         const tx = sendNft(api, "sent", attacker, collectionId, childNftId, newOwnerNFT);
 
@@ -115,7 +112,7 @@ describe("Integration test: send NFT", () => {
         expect(isChild).to.be.false;
     });
 
-    it("[Negative] send NFT to non-existing NFT", async () => {
+    it("[negative] send NFT to non-existing NFT", async () => {
         const originalOwnerUri = alice;
 
         const collectionId = await createTestCollection(alice);
@@ -123,7 +120,7 @@ describe("Integration test: send NFT", () => {
         const parentNftId = maxNftId;
         const childNftId = await mintNft(api, alice, originalOwnerUri, collectionId, "child-nft-metadata");
 
-        const newOwnerNFT: Nft = [collectionId, parentNftId];
+        const newOwnerNFT: NftIdTuple = [collectionId, parentNftId];
 
         const tx = sendNft(api, "sent", alice, collectionId, childNftId, newOwnerNFT);
 
@@ -157,7 +154,7 @@ describe("Integration test: send NFT", () => {
         const parentNftId = await mintNft(api, alice, ownerAlice, aliceCollectionId, "parent-nft-metadata");
         const childNftId = await mintNft(api, bob, ownerBob, bobCollectionId, "child-nft-metadata");
 
-        const newOwnerNFT: Nft = [aliceCollectionId, parentNftId];
+        const newOwnerNFT: NftIdTuple = [aliceCollectionId, parentNftId];
 
         await sendNft(api, "pending", bob, bobCollectionId, childNftId, newOwnerNFT);
 
@@ -165,13 +162,13 @@ describe("Integration test: send NFT", () => {
         expect(isChild).to.be.false;
     });
 
-    it("[Negative] unable to send NFT to itself", async () => {
+    it("[negative] unable to send NFT to itself", async () => {
         const nftOwner = alice;
         const collectionId = await createTestCollection(alice);
 
         const nftId = await mintNft(api, alice, nftOwner, collectionId, "ouroboros-nft-metadata");
 
-        const newOwnerNFT: Nft = [collectionId, nftId];
+        const newOwnerNFT: NftIdTuple = [collectionId, nftId];
 
         const tx = sendNft(api, "sent", alice, collectionId, nftId, newOwnerNFT);
 
@@ -181,7 +178,7 @@ describe("Integration test: send NFT", () => {
         expect(isChild).to.be.false;
     });
 
-    it("[Negative] unable to send NFT to child NFT", async () => {
+    it("[negative] unable to send NFT to child NFT", async () => {
         const originalOwnerUri = alice;
 
         const collectionId = await createTestCollection(alice);
@@ -189,14 +186,14 @@ describe("Integration test: send NFT", () => {
         const parentNftId = await mintNft(api, alice, originalOwnerUri, collectionId, "parent-nft-metadata");
         const childNftId = await mintNft(api, alice, originalOwnerUri, collectionId, "child-nft-metadata");
 
-        const newOwnerNFT: Nft = [collectionId, parentNftId];
+        const newOwnerNFT: NftIdTuple = [collectionId, parentNftId];
 
         await sendNft(api, "sent", alice, collectionId, childNftId, newOwnerNFT);
 
         const isChild = await isNftChildOfAnother(api, collectionId, childNftId, newOwnerNFT);
         expect(isChild).to.be.true;
 
-        const descendentOwner: Nft = [collectionId, childNftId];
+        const descendentOwner: NftIdTuple = [collectionId, childNftId];
         const tx = sendNft(api, "sent", alice, collectionId, parentNftId, descendentOwner);
 
         await expectTxFailure(/CannotSendToDescendentOrSelf/, tx);
@@ -204,7 +201,7 @@ describe("Integration test: send NFT", () => {
         expect(isOuroboros).to.be.false;
     });
 
-    it("[Negative] unable to send NFT to descendent NFT", async () => {
+    it("[negative] unable to send NFT to descendent NFT", async () => {
         const originalOwnerUri = alice;
 
         const collectionId = await createTestCollection(alice);
@@ -213,20 +210,20 @@ describe("Integration test: send NFT", () => {
         const childNftId = await mintNft(api, alice, originalOwnerUri, collectionId, "child-nft-metadata");
         const grandsonNftId = await mintNft(api, alice, originalOwnerUri, collectionId, "grandson-nft-metadata");
 
-        const ownerParentNFT: Nft = [collectionId, parentNftId];
+        const ownerParentNFT: NftIdTuple = [collectionId, parentNftId];
 
         await sendNft(api, "sent", alice, collectionId, childNftId, ownerParentNFT);
 
         const isChild = await isNftChildOfAnother(api, collectionId, childNftId, ownerParentNFT);
         expect(isChild).to.be.true;
 
-        const ownerChildNFT: Nft = [collectionId, childNftId];
+        const ownerChildNFT: NftIdTuple = [collectionId, childNftId];
         await sendNft(api, "sent", alice, collectionId, grandsonNftId, ownerChildNFT);
 
         const isGrandson = await isNftChildOfAnother(api, collectionId, grandsonNftId, ownerChildNFT);
         expect(isGrandson).to.be.true;
 
-        const ownerGrandsonNFT: Nft = [collectionId, grandsonNftId];
+        const ownerGrandsonNFT: NftIdTuple = [collectionId, grandsonNftId];
         const tx = sendNft(api, "sent", alice, collectionId, parentNftId, ownerGrandsonNFT);
 
         await expectTxFailure(/CannotSendToDescendentOrSelf/, tx);
