--- tests/src/util/getCollectionsCount.ts
@@ -1,6 +1,0 @@
-import {ApiPromise} from '@polkadot/api';
-import '../interfaces/augment-api-query';
-
-export async function getCollectionsCount(api: ApiPromise): Promise<number> {
-    return (await api.query.rmrkCore.collectionIndex()).toNumber();
-}
