+++ tests/src/util/helpers.ts
@@ -1,10 +1,45 @@
-import type {EventRecord} from '@polkadot/types/interfaces';
-import type {GenericEventData} from '@polkadot/types';
-import { expect } from 'chai';
+import { ApiPromise } from "@polkadot/api";
+import { RmrkTraitsNftAccountIdOrCollectionNftTuple as NftOwner } from "@polkadot/types/lookup";
+import type { EventRecord } from '@polkadot/types/interfaces';
+import type { GenericEventData } from '@polkadot/types';
+import privateKey from "../substrate/privateKey";
+import { NftIdTuple, getChildren } from './fetch';
+import chaiAsPromised from 'chai-as-promised';
+import chai from 'chai';
+
+chai.use(chaiAsPromised);
+const expect = chai.expect;
 
 interface TxResult<T> {
     success: boolean;
     successData: T | null;
+}
+
+export function makeNftOwner(api: ApiPromise, owner: string | NftIdTuple): NftOwner {
+    const isNftSending = (typeof owner !== "string");
+
+    if (isNftSending) {
+        return api.createType("RmrkTraitsNftAccountIdOrCollectionNftTuple", {
+            "CollectionAndNftTuple": owner
+        });
+    } else {
+        return api.createType("RmrkTraitsNftAccountIdOrCollectionNftTuple", {
+            "AccountId": privateKey(owner).address
+        });
+    }
+}
+
+export async function isNftChildOfAnother(
+    api: ApiPromise,
+    collectionId: number,
+    nftId: number,
+    parentNft: NftIdTuple
+): Promise<boolean> {
+    return (await getChildren(api, parentNft[0], parentNft[1]))
+        .find((child) => {
+            return child[0] === collectionId
+                && child[1] === nftId;
+        }) !== undefined;
 }
 
 export function isTxResultSuccess(events: EventRecord[]): boolean {
